Exclude broken test project
Trying to run the instrumented tests in the domain project results in
the runner crashing due to being unable to find all required classes.

It's possible to introduce this error in any project just by depending
on tox4j-c.

It is still possible to run the instrumented tests for the domain
project locally if you explicitly run the test class instead of relying
on the AndroidJUnitRunner to find all tests.