Streamline file transfer handling
Now, instead of prompting the user for a storage location before
accepting, the file transfer is accepted into an internal area, and an
option to allow the user to export the file to a user-provided area was
added.

This commit also fixes the whole tapping an image opening the gallery
thing. It was broken before.