Fix it being possible to get some settings stuck out of view
This could only happen in landscape mode, and only if the view was
recreated while you were on it. No idea why Android decides to change
the size of it when recreating it, but here we are. If I think of a
nicer solution I'll revert this and fix it properly.