Fix rare crash on canceling a file transfer
At first I thought this was a lot more common than it was, but in
addition to what's in the title, you need to be transfering the same
file in multiple transfers, and have e.g. closed aTox and reopened it
from a notification or have changed phone orientation in order to
trigger the crash.