Update the Tox test to not require mockk
This required some previously internal repository constructors to be
made public, but it's worth it to drop a dependency.

This commit also removes the now-unused mockk dependency. We've been
unable to update it for a very long time as the versions after the one
we're using only supports API 21 and newer, and dropping support for API
19 just for a mocking library doesn't seem reasonable.