Disable broken InvalidPackageDeclaration rule in detekt
We set a package prefix in the Gradle files so we don't have to have
that unreasonably deep folder structure most Gradle projects do, and
detekt doesn't seem to understand that. There are some configuration
options to do with ignoring a part of a package, but setting that up in
a multi-module project seems like too much work when ktlint already
checks this.