Update protobuf, libsodium, opus, and libvpx
All except protobuf were updated to the latest versions.

Protobuf was updated to the latest version with support for autotools
and will be migrated to being built with CMake later. Not updating
protobuf to the latest version shouldn't be a big deal as it's only used
internally in tox4j and is not used for serialization of data from the
network.