Fix possible race in user creation
This issue showed up a lot in CI where things are a lot slower. The
createprofile fragment spawns a coroutine that will create a user and
returns before it's done, then the contactlist fragment will try to load
the user, leading to a crash if the user-creation coroutine hasn't done
its thing yet.

With this change, we wait for the user-creation coroutine to complete
before moving to the contactlist fragment.

To reproduce the issue locally, I added a delay(1000) in
UserManager::create.